#1602 groundwork for adding a SOAP style client
