+++ tools/requirements.txt
@@ -4,3 +4,4 @@ requests >= 2.20.0
 prompt_toolkit >= 2
 pygments >= 2.0.0
 urllib3 >= 1.24
+zeep
