+++ tools/test-requirements.txt
@@ -10,3 +10,4 @@ requests >= 2.20.0
 prompt_toolkit >= 2
 pygments >= 2.0.0
 urllib3 >= 1.24
+zeep
